+++ src/api.js
@@ -1,9 +1,11 @@
 // api.js provides the functions to request the reddit data
 
 const axios = require("axios");
+const vscode = require("vscode");
+let config = vscode.workspace.getConfiguration("redditviewer");
 
 const web = axios.create({
-  timeout: 2000
+  timeout: config.requestTimeout
 });
 
 // requests the subreddit data
@@ -22,7 +24,7 @@ function getSubreddit(subreddit, sort, interval, limit, count, after, before) {
         resolve(response.data);
       })
       .catch(error => {
-        reject(error);
+        reject(error.message);
       });
   });
 }
@@ -36,7 +38,7 @@ function getArticle(subreddit, article) {
         resolve(response.data);
       })
       .catch(error => {
-        reject(error);
+        reject(error.message);
       });
   });
 }
@@ -50,7 +52,111 @@ function getTrendingSubreddits() {
         resolve(response.data.subreddit_names);
       })
       .catch(error => {
-        reject(error);
+        reject(error.message);
+      });
+  });
+}
+
+// requests the given view of an user
+// can be 'about', 'posts', 'comments'
+function getUser(username, view) {
+  return new Promise((resolve, reject) => {
+    let url;
+    if (view === "posts") {
+      url = `https://reddit.com/user/${username}/submitted.json`;
+    } else {
+      url = `https://reddit.com/user/${username}/${view}.json`;
+    }
+    web
+      .get(url)
+      .then(response => {
+        resolve({
+          view: view,
+          data: response.data
+        });
+      })
+      .catch(error => {
+        reject(error.message);
+      });
+  });
+}
+
+// get the collections of the current user
+function getCollections(cookie) {
+  return new Promise((resolve, reject) => {
+    web
+      .get(`https://www.reddit.com/api/multi/mine`, {
+        headers: {
+          Cookie: `reddit_session=${encodeURIComponent(cookie)}`
+        }
+      })
+      .then(response => {
+        resolve(response.data);
+      })
+      .catch(error => {
+        reject(error.message);
+      });
+  });
+}
+
+// get the collection with the given path
+function getCollection(path, cookie) {
+  return new Promise((resolve, reject) => {
+    web
+      .get(`https://www.reddit.com${path}.json`, {
+        headers: {
+          Cookie: `reddit_session=${encodeURIComponent(cookie)}`
+        }
+      })
+      .then(response => {
+        resolve(response.data);
+      })
+      .catch(error => {
+        reject(error.message);
+      });
+  });
+}
+
+// request a user login with given username and password
+function userLogin(username, password) {
+  return new Promise((resolve, reject) => {
+    // create login request with encoded password
+    let url = `https://ssl.reddit.com/api/login/${username}?api_type=json&user=${username}&passwd=${encodeURIComponent(
+      password
+    )}`;
+    web
+      .post(url)
+      .then(response => {
+        if (response.data.json.data !== undefined) {
+          resolve(response.data.json.data.cookie);
+        } else {
+          reject();
+        }
+      })
+      .catch(() => {
+        reject();
+      });
+  });
+}
+
+// checks if current session is valid
+function checkSession(cookie) {
+  return new Promise((resolve, reject) => {
+    web
+      .get(`https://reddit.com/.json`, {
+        headers: {
+          Cookie: `reddit_session=${encodeURIComponent(cookie)}`
+        }
+      })
+      .then(response => {
+        if (response.data.data.modhash) {
+          resolve();
+        } else {
+          reject();
+        }
+      })
+      .catch(() => {
+        reject();
       });
   });
 }
@@ -58,5 +164,10 @@ function getTrendingSubreddits() {
 module.exports = {
   getSubreddit,
   getArticle,
-  getTrendingSubreddits
+  getTrendingSubreddits,
+  getUser,
+  getCollections,
+  getCollection,
+  userLogin,
+  checkSession
 };